Tracking: Move tracking into a submodule
The tracking is becoming increasingly complex so move into a new
submodule to allow splitting but keeping logical cohesion.
