+++ kano_profile/tracker/__init__.py
@@ -1,6 +1,5 @@
-#!/usr/bin/env python
-
-# tracker.py
+#
+# __init__.py
 #
 # Copyright (C) 2014 - 2017 Kano Computing Ltd.
 # License: http://www.gnu.org/licenses/gpl-2.0.txt GNU GPL v2
